When building with Qt, use Qt image loaders
This reduces dependencies: no more libjpeg or libpng needed. They are
still available via -DIMAGES=SYSTEM if you'd prefer to use them.
