+++ garglk/CMakeLists.txt
@@ -24,10 +24,25 @@ endif()
 set(WITH_TTS "AUTO" CACHE STRING "Enable text-to-speech support (ON/OFF/AUTO/DYNAMIC)")
 set(GARGLKPRE "" CACHE STRING "Binary prefix")
 set(SOUND "SDL" CACHE STRING "Backend to use for sound (SDL, QT, or NONE)")
-set(JPEGLIB "AUTO" CACHE STRING "JPEG library to use (TURBO, IJG, or AUTO)")
+if(INTERFACE STREQUAL "QT")
+    set(IMAGES_DEFAULT "QT")
+else()
+    set(IMAGES_DEFAULT "SYSTEM")
+endif()
+set(IMAGES "${IMAGES_DEFAULT}" CACHE STRING "Backend to use for image loading (SYSTEM {for system libraries} or QT)")
+set_property(CACHE IMAGES PROPERTY STRINGS SYSTEM QT)
+set(JPEGLIB "AUTO" CACHE STRING "JPEG library to use (TURBO, IJG, or AUTO); ignored if IMAGES set to QT")
 
 if(INTERFACE STREQUAL "QT")
     option(WITH_NATIVE_FILE_DIALOGS "Use native dialogs instead of Qt dialogs" ON)
+endif()
+
+if(NOT IMAGES STREQUAL "SYSTEM" AND NOT IMAGES STREQUAL "QT")
+    message(FATAL_ERROR "IMAGES must be either SYSTEM or QT")
+endif()
+
+if(IMAGES STREQUAL "QT" AND NOT INTERFACE STREQUAL "QT")
+    message(FATAL_ERROR "IMAGES=QT requires INTERFACE=QT")
 endif()
 
 if(SOUND STREQUAL "SDL" OR SOUND STREQUAL "QT")
@@ -79,8 +94,14 @@ if(INTERFACE STREQUAL "QT")
     endif()
 endif()
 
+if(IMAGES STREQUAL "QT")
+    set(IMGLOAD imgload-qt.cpp)
+else()
+    set(IMGLOAD imgload-system.cpp)
+endif()
+
 add_library(garglk-common OBJECT babeldata.cpp style.cpp config.cpp draw.cpp event.cpp
-    garglk.cpp imgload.cpp imgscale.cpp theme.cpp winblank.cpp window.cpp
+    garglk.cpp imgload.cpp ${IMGLOAD} imgscale.cpp theme.cpp winblank.cpp window.cpp
     wingfx.cpp wingrid.cpp winmask.cpp winpair.cpp wintext.cpp zbleep.cpp
     ${GARGLKINI_CXX} ${THEME_DARK_CXX} ${THEME_LIGHT_CXX}
 
@@ -125,9 +146,8 @@ else()
 endif()
 
 find_package(Freetype REQUIRED)
-find_package(PNG 1.6 REQUIRED)
-target_include_directories(garglk-common PUBLIC cheapglk PRIVATE ${FREETYPE_INCLUDE_DIRS} ${PNG_INCLUDE_DIRS})
-target_link_libraries(garglk-common PRIVATE ${FREETYPE_LIBRARIES} ${PNG_LIBRARIES})
+target_include_directories(garglk-common PUBLIC cheapglk PRIVATE ${FREETYPE_INCLUDE_DIRS})
+target_link_libraries(garglk-common PRIVATE ${FREETYPE_LIBRARIES})
 
 if(NOT WITH_BUNDLED_FMT)
     find_package(fmt CONFIG)
@@ -274,41 +294,47 @@ else()
     target_compile_definitions(garglk-common PRIVATE GARGLK_CONFIG_NO_SOUND)
 endif()
 
-if(NOT JPEGLIB STREQUAL "AUTO")
-    set(JPEG_REQUIRED "REQUIRED")
-endif()
+if(IMAGES STREQUAL "SYSTEM")
+    find_package(PNG 1.6 REQUIRED)
+    target_include_directories(garglk-common PRIVATE ${PNG_INCLUDE_DIRS})
+    target_link_libraries(garglk-common PRIVATE ${PNG_LIBRARIES})
 
-if(JPEGLIB STREQUAL "AUTO" OR JPEGLIB STREQUAL "TURBO")
-    find_package(PkgConfig ${JPEG_REQUIRED})
-    if(PkgConfig_FOUND)
-        pkg_check_modules(TURBO ${JPEG_REQUIRED} IMPORTED_TARGET libturbojpeg>=2.0)
-        if(TARGET PkgConfig::TURBO)
-            target_include_directories(garglk-common PRIVATE PkgConfig::TURBO)
-            target_link_libraries(garglk-common PRIVATE PkgConfig::TURBO)
-            target_compile_definitions(garglk-common PRIVATE GARGLK_CONFIG_JPEG_TURBO)
-            set(GARGLK_FOUND_JPEG TRUE)
+    if(NOT JPEGLIB STREQUAL "AUTO")
+        set(JPEG_REQUIRED "REQUIRED")
+    endif()
+
+    if(JPEGLIB STREQUAL "AUTO" OR JPEGLIB STREQUAL "TURBO")
+        find_package(PkgConfig ${JPEG_REQUIRED})
+        if(PkgConfig_FOUND)
+            pkg_check_modules(TURBO ${JPEG_REQUIRED} IMPORTED_TARGET libturbojpeg>=2.0)
+            if(TARGET PkgConfig::TURBO)
+                target_include_directories(garglk-common PRIVATE PkgConfig::TURBO)
+                target_link_libraries(garglk-common PRIVATE PkgConfig::TURBO)
+                target_compile_definitions(garglk-common PRIVATE GARGLK_CONFIG_JPEG_TURBO)
+                set(GARGLK_FOUND_JPEG TRUE)
+            endif()
+        else()  # e.g. Windows-on-ARM cross-compilation
+            find_library(TURBO NAMES turbojpeg libturbojpeg)
+            if(NOT "${TURBO}" STREQUAL "TURBO-NOTFOUND")
+                target_link_libraries(garglk-common PRIVATE ${TURBO})
+                target_compile_definitions(garglk-common PRIVATE GARGLK_CONFIG_JPEG_TURBO)
+                set(GARGLK_FOUND_JPEG TRUE)
+            endif()
         endif()
-    else()  # e.g. Windows-on-ARM cross-compilation
-        find_library(TURBO NAMES turbojpeg libturbojpeg)
-        if(NOT "${TURBO}" STREQUAL "TURBO-NOTFOUND")
-            target_link_libraries(garglk-common PRIVATE ${TURBO})
-            target_compile_definitions(garglk-common PRIVATE GARGLK_CONFIG_JPEG_TURBO)
+    endif()
+
+    if(NOT GARGLK_FOUND_JPEG AND (JPEGLIB STREQUAL "AUTO" OR JPEGLIB STREQUAL "IJG"))
+        find_package(JPEG 80 ${JPEG_REQUIRED})
+        if(JPEG_FOUND)
+            target_include_directories(garglk-common PRIVATE "${JPEG_INCLUDE_DIRS}")
+            target_link_libraries(garglk-common PRIVATE "${JPEG_LIBRARIES}")
             set(GARGLK_FOUND_JPEG TRUE)
         endif()
     endif()
-endif()
 
-if(NOT GARGLK_FOUND_JPEG AND (JPEGLIB STREQUAL "AUTO" OR JPEGLIB STREQUAL "IJG"))
-    find_package(JPEG 80 ${JPEG_REQUIRED})
-    if(JPEG_FOUND)
-        target_include_directories(garglk-common PRIVATE "${JPEG_INCLUDE_DIRS}")
-        target_link_libraries(garglk-common PRIVATE "${JPEG_LIBRARIES}")
-        set(GARGLK_FOUND_JPEG TRUE)
+    if(NOT GARGLK_FOUND_JPEG)
+        message(FATAL_ERROR "Unable to find a usable JPEG library")
     endif()
-endif()
-
-if(NOT GARGLK_FOUND_JPEG)
-    message(FATAL_ERROR "Unable to find a usable JPEG library")
 endif()
 
 target_compile_definitions(garglk-common PRIVATE GARGLK_CONFIG_SCALERS)
